+++ otherlibs/win32unix/unix.ml
@@ -296,20 +296,15 @@ module LargeFile =
 
 (* Mapping files into memory *)
 
-type map_file_impl =
-  { map_file_impl
-    : 'a 'b 'c. file_descr
-      -> ('a, 'b) CamlinternalBigarray.kind
-      -> 'c CamlinternalBigarray.layout
-      -> bool
-      -> int array
-      -> int64
-      -> ('a, 'b, 'c) CamlinternalBigarray.genarray
-  }
-let map_file_impl =
-  ref { map_file_impl = fun _ _ _ _ _ _ -> failwith "Bigarray not initialized!" }
+external map_internal:
+   file_descr -> ('a, 'b) CamlinternalBigarray.kind
+              -> 'c CamlinternalBigarray.layout
+              -> bool -> int array -> int64
+              -> ('a, 'b, 'c) CamlinternalBigarray.genarray
+     = "caml_unix_map_file_bytecode" "caml_unix_map_file"
+
 let map_file fd ?(pos=0L) kind layout shared dims =
-  !map_file_impl.map_file_impl fd kind layout shared dims pos
+  map_internal fd kind layout shared dims pos
 
 (* File permissions and ownership *)
